progression-command/progression-command-api/src/test/java/uk/gov/moj/cpp/progression/command/RemoveCaseFromGroupCasesApiTest.java
package uk.gov.moj.cpp.progression.command;

import static java.util.UUID.randomUUID;
import static javax.json.Json.createObjectBuilder;
import static org.hamcrest.MatcherAssert.assertThat;
import static org.hamcrest.Matchers.is;
import static org.mockito.Mockito.verify;

import uk.gov.justice.services.core.sender.Sender;
import uk.gov.justice.services.messaging.Envelope;
import uk.gov.justice.services.messaging.JsonEnvelope;
import uk.gov.justice.services.messaging.Metadata;
import uk.gov.justice.services.messaging.spi.DefaultEnvelope;
import uk.gov.justice.services.messaging.spi.DefaultJsonEnvelopeProvider;

import javax.json.JsonObject;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.ArgumentCaptor;
import org.mockito.Captor;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.runners.MockitoJUnitRunner;


@RunWith(MockitoJUnitRunner.class)
public class RemoveCaseFromGroupCasesApiTest {

    @Mock
    private Sender sender;

    @InjectMocks
    private RemoveCaseFromGroupCasesApi removeCaseFromGroupCasesApi;

    @Captor
    private ArgumentCaptor<DefaultEnvelope> envelopeCaptor;

    @Test
    public void shouldHandleRemoveCaseFromGroupCases() {

        final JsonObject payload = createObjectBuilder()
                .add("prosecutionCaseId", randomUUID().toString())
                .add("groupId", randomUUID().toString())
                .build();

        final Metadata metadata = Envelope.metadataBuilder()
                .withName("progression.remove-case-from-group-cases")
                .withId(randomUUID())
                .withUserId(randomUUID().toString())
                .build();

        final JsonEnvelope commandEnvelope = new DefaultJsonEnvelopeProvider().envelopeFrom(metadata, payload);

        removeCaseFromGroupCasesApi.handle(commandEnvelope);

        verify(sender).send(envelopeCaptor.capture());

        final DefaultEnvelope capturedEnvelope = envelopeCaptor.getValue();
        assertThat(capturedEnvelope.metadata().name(), is("progression.command.remove-case-from-group-cases"));
        assertThat(capturedEnvelope.payload(), is(payload));
    }
}
